% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medRCT_shiny.R
\name{catch_model_messages}
\alias{catch_model_messages}
\title{Capture Warnings and Errors from Model Fitting}
\usage{
catch_model_messages(formula, data, family)
}
\arguments{
\item{formula}{A formula specifying the model to be fitted.}

\item{data}{A data frame containing the variables referenced in the formula.}

\item{family}{A description of the error distribution and link function to
be used in the model, as specified in \code{glm} (it can be either \code{binomial} or \code{gaussian}).}
}
\description{
\code{catch_model_messages} fits a generalised linear model while capturing
and storing any warnings or errors generated during the fitting process.
This is useful for debugging.
}
\details{
This function uses \code{tryCatch} and \code{withCallingHandlers} to handle warnings
and errors separately. Warnings are captured and stored in the resulting
model object under the \code{warnings} attribute, while errors are stored under
the \code{errors} attribute. If the model fits successfully without warnings or errors,
these attributes will be \code{NULL}.
}
\keyword{internal}
