% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cf_predict}
\alias{cf_predict}
\title{Counterfactual Prediction and Random Draw}
\usage{
cf_predict(fit, data, var_name, n, family)
}
\arguments{
\item{fit}{A fitted glm model object}

\item{data}{A \code{data.table} containing the data to which the counterfactual predictions
will be applied.}

\item{var_name}{A character string specifying the name of the variable to store
the random draws.}

\item{n}{An integer specifying the number of observations in the dataset.}

\item{family}{A character string specifying the distribution family to use for
generating random draws. Must be either \code{"binomial"} or \code{"gaussian"}.}
}
\description{
Counterfactual Prediction and Random Draw
}
