% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medRCT_shiny.R
\name{collect_models}
\alias{collect_models}
\title{Collect All Models fitted by the algorithm}
\usage{
collect_models(
  data,
  exposure,
  outcome,
  mediators,
  intermediate_confs,
  confounders,
  interactions_XC = "all",
  intervention_type = c("all", "shift_all", "shift_k", "shift_k_order")
)
}
\arguments{
\item{data}{A \code{data.frame} containing the dataset for analysis.}

\item{exposure}{A \code{character} string specifying the name of the exposure variable in the dataset.
The exposure variable can be categorical, with \code{0} explicitly denoting the unexposed (or control) group.
Other values represent different exposure categories.}

\item{outcome}{A \code{character} string specifying the name of the outcome variable in the dataset.
The outcome variable can be specified as either binary or continuous.}

\item{mediators}{A \code{character} vector specifying the names of mediator variables in the dataset. The mediators
can be specified as either binary or continuous. When estimating the effect type \code{"shift_k_order"},
the order of mediators in the vector is important, as it determines the causal sequence of mediators.}

\item{intermediate_confs}{A \code{character} vector specifying the names of intermediate confounders in the dataset.
The intermediate confounders can be specified as either binary or continuous. If \code{NULL},
no intermediate confounders are specified, and the natural effect will be estimated.}

\item{confounders}{A \code{character} vector listing the names of baseline confounders.}

\item{interactions_XC}{A \code{character} string specifying the exposure-confounder or confounder-confounder
interaction terms to include in the regression models for confounder adjustment. The default value, \code{"all"},
includes all two-way exposure-confounder interactions but excludes confounder-confounder interactions.
Specify \code{"none"} to exclude all two-way exposure-confounder and confounder-confounder interactions.}

\item{intervention_type}{A \code{character} string indicating the type of interventional effect to be estimated.}
}
\description{
\code{collect_models} fits and collects models required for the algorithm.
}
\keyword{internal}
