% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{family_type}
\alias{family_type}
\title{Determine Family Type for GLM}
\usage{
family_type(data, variable_names, unique_threshold = 10)
}
\arguments{
\item{data}{A data frame containing the variables to be analyzed.}

\item{variable_names}{A character vector of variable names to check. Each variable name should correspond
to a column in the data frame.}

\item{unique_threshold}{An integer value specifying the minimum number of unique values for a variable to
be considered continuous. Default is 10.}
}
\value{
A list where each element corresponds to a family type for variables from \code{variable_names}
and the value is either "binomial()" or "gaussian()", indicating the type of the family.
}
\description{
Determine Family Type for GLM
}
