% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fit_model}
\alias{fit_model}
\title{Fit a generalized linear model with separation handling}
\usage{
fit_model(formula, data, family, separation_method, ...)
}
\arguments{
\item{formula}{A formula specifying the model.}

\item{data}{A data frame containing the variables in the model.}

\item{family}{Link function to be used in the model.}

\item{separation_method}{Method to handle separation, only relevant for binomial (binary outcome) models.
Options are \code{"brglm"} (fits a bias-reduced GLM using the \code{brglm2} package)
or \code{"discard"} (if separation is detected using the \code{detectseparation} package, the function returns \code{NA}
and the bootstrap sample is discarded).}

\item{...}{Additional arguments passed to \code{glm}.}
}
\value{
An object of class \code{glm} or \code{brglmFit}, depending on the method used.
}
\description{
This function fits a generalized linear model and handles data separation in binomial models
using either the \code{brglm2} package for bias-reduced estimation or the \code{detectseparation}
method to discard separated data.
}
\keyword{internal}
