% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medRCT_shiny.R
\name{flatten_models}
\alias{flatten_models}
\title{Flatten a Nested Model List
This function processes a nested list of models by extracting specific sublists (\code{L_models} and \code{M_models}),
renaming their elements, and merging them into the parent level.}
\usage{
flatten_models(model_list)
}
\arguments{
\item{model_list}{A nested list containing models, with potential sublists named \code{L_models} and \code{M_models}.}
}
\description{
Flatten a Nested Model List
This function processes a nested list of models by extracting specific sublists (\code{L_models} and \code{M_models}),
renaming their elements, and merging them into the parent level.
}
\keyword{internal}
