% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medRCT_shiny.R
\name{gen_formula_shiny}
\alias{gen_formula_shiny}
\title{Generate Model Formulas for Shiny-Based Model Assessment}
\usage{
gen_formula_shiny(
  mediators = mediators,
  exposure = exposure,
  k,
  first = NULL,
  MM = NULL,
  K = NULL,
  interactions_XC,
  include_all = FALSE,
  marginal = FALSE
)
}
\arguments{
\item{mediators}{A \code{character} vector specifying the names of the variables in the dataset corresponding to the mediators of interest. The mediators
can be either binary or continuous. When estimating the effect type \code{"shift_k_order"},
the order of mediators in the vector is important, as it interpreted as the assumed causal ordering of the mediators.}

\item{exposure}{A \code{character} string specifying the name of the exposure variable in the dataset.
The exposure variable must be categorical, with \code{0} explicitly denoting the unexposed (or control) group, which is taken as the reference group.
Other values represent different, non-reference exposure categories.}

\item{k}{An integer specifying the index of the mediator for which the formula is being generated.}

\item{first}{An integer (optional) specifying the index of the first mediator. Defaults to \code{NULL},
indicating that the function will generate formulas without explicitly considering this parameter.}

\item{MM}{An integer (optional) specifying the index of the mediator whose distribution will be shifted.
Defaults to \code{NULL}, indicating that the function will generate formulas without explicitly considering
this parameter.}

\item{K}{An integer (optional) specifying the total number of mediators and intermediate confounders.
Defaults to \code{NULL}, indicating that the function will generate formulas without explicitly considering this
parameter.}

\item{interactions_XC}{A \code{character} string specifying the two-way interactions amongst exposure and baseline confounders
to include in the regression models in the estimation procedure. The default value, \code{"all"},
includes all two-way exposure-confounder interactions but excludes confounder-confounder interactions.
Specify \code{"none"} to exclude all two-way interactions amongst exposure and baseline confounders.}

\item{include_all}{Logical.}

\item{marginal}{Logical. If \code{TRUE}, estimating marginals under \code{X=0}.}
}
\description{
This function generates model formulas using the original variable names as they appear in the dataset.
}
\keyword{internal}
