% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_proc.R
\name{joint_dist}
\alias{joint_dist}
\title{Estimation of Joint Distribution of Mediators and Intermediate Confounders under
Exposed and Unexposed}
\usage{
joint_dist(
  k,
  K,
  data,
  dat2,
  fam_type,
  mediators,
  interactions_XC,
  exposure_level,
  n
)
}
\arguments{
\item{k}{An integer specifying the index of the current mediator being processed.}

\item{K}{An integer specifying the total number of mediators and intermediate confounders.}

\item{data}{A \code{data.frame} containing the dataset for analysis. The dataset is
used to fit the model for the mediators.}

\item{dat2}{A \code{data.frame} used to perform counterfactual predictions and
random draws.}

\item{fam_type}{A \code{character} string specifying the family type for modeling.
Options typically include \code{"gaussian"} for continuous variables or \code{"binomial"}
for binary variables.}

\item{mediators}{A \code{character} vector including the variable names for mediators
(including intermediate confounders).}

\item{interactions_XC}{A \code{character} string specifying the two-way interactions amongst exposure and baseline confounders
to include in the regression models in the estimation procedure.}

\item{exposure_level}{A numeric vector specifying the levels of the exposure
(e.g., \code{c(0, 1)}) for which counterfactual predictions are performed.}

\item{n}{An integer specifying the number of observations for \code{dat2}.}
}
\description{
Estimation of Joint Distribution of Mediators and Intermediate Confounders under
Exposed and Unexposed
}
\keyword{internal}
