% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_proc.R
\name{joint_unexposed}
\alias{joint_unexposed}
\title{Estimation of Joint Distribution of All Mediators under Unexposed}
\usage{
joint_unexposed(
  k,
  first,
  K,
  data,
  dat2,
  fam_type,
  mediators,
  interactions_XC,
  separation_method,
  n
)
}
\arguments{
\item{k}{An integer specifying the index of the current mediator being processed.}

\item{first}{An integer specifying the index of the first mediator.}

\item{K}{An integer specifying the total number of mediators and intermediate confounders.}

\item{data}{A \code{data.frame} containing the dataset for analysis. The dataset is
used to fit the model for the mediators.}

\item{dat2}{A \code{data.frame} used to perform counterfactual predictions and
random draws.}

\item{fam_type}{A \code{character} string specifying the family type for modeling.
Options typically include \code{"gaussian"} for continuous variables or \code{"binomial"}
for binary variables.}

\item{mediators}{A \code{character} vector including the variable names for mediators
(including intermediate confounders).}

\item{interactions_XC}{A \code{character} string specifying the two-way interactions amongst exposure and baseline confounders
to include in the regression models in the estimation procedure.}

\item{separation_method}{Method to handle separation, only relevant for binomial (binary outcome) models.
Options are \code{"brglm"} (Logistic regression models are fitted using bias reduction methods for generalised linear models implemented in the \code{brglm2} package)
or \code{"discard"} (if separation is detected, the function returns \code{NA}. If this occurs during the main estimation,
the program stops; if it occurs during bootstrapping, the affected bootstrap samples are discarded).}

\item{n}{An integer specifying the number of observations for \code{dat2}.}
}
\description{
Estimation of Joint Distribution of All Mediators under Unexposed
}
\keyword{internal}
