% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_proc.R
\name{joint_X_nonzero}
\alias{joint_X_nonzero}
\title{Estimation of Joint Distribution of Mediators under Exposed}
\usage{
joint_X_nonzero(
  MM,
  k,
  first,
  K,
  data,
  dat2,
  fam_type,
  interactions_XC,
  lnzero,
  n,
  index
)
}
\arguments{
\item{MM}{An integer specifying the index of the mediator whose distribution
will be shifted.}

\item{k}{An integer specifying the index of the current mediator being processed.}

\item{first}{An integer specifying the index of the first mediator.}

\item{K}{An integer specifying the total number of mediators and intermediate confounders.}

\item{data}{A \code{data.frame} containing the dataset for analysis. The dataset is
used to fit the model for the mediators.}

\item{dat2}{A \code{data.frame} used to perform counterfactual predictions and
random draws.}

\item{fam_type}{A \code{character} string specifying the family type for modeling.
Options typically include \code{"gaussian"} for continuous variables or \code{"binomial"}
for binary variables.}

\item{interactions_XC}{A \code{character} string specifying the exposure-confounder or
confounder-confounder interaction terms to include in the regression models for
confounder adjustment.}

\item{lnzero}{A numeric vector specifying the non-zero levels of the exposure.}

\item{n}{An integer specifying the number of observations for \code{dat2}.}

\item{index}{An integer vector specifying the indices of all mediators, excluding the mediator
specified by \code{MM}.}
}
\description{
Estimation of Joint Distribution of Mediators under Exposed
}
