% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LSAC.R
\docType{data}
\name{LSACdata}
\alias{LSACdata}
\title{The Longitudinal Study of Australian Children data}
\format{
A data frame containing following variables from a longitudinal cohort study.
\describe{
\item{sep}{A binary exposure variable}
\item{fam_stress}{A binary intermediate confounder}
\item{parent_mh}{A binary mediator}
\item{preschool_att}{A binary mediator}
\item{child_mh}{A binary outcome variable}
\item{child_sex:}{A binary baseline confounder}
\item{child_atsi}{A binary baseline confounder}
\item{mat_cob}{A binary baseline confounder}
\item{mat_engl}{A binary baseline confounder}
\item{mat_age}{A binary baseline confounder}
}
}
\usage{
LSACdata
}
\description{
Simulated data based on the Longitudinal Study of Australian Children containing 5,107 rows and
10 columns with no missing values.
}
\keyword{datasets}
