% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_fun.R
\name{medRCT.fun}
\alias{medRCT.fun}
\title{Causal Mediation Analysis Estimating Interventional Effects}
\usage{
medRCT.fun(
  dat,
  ind = 1:nrow(dat),
  first,
  K,
  fam_type,
  mediators,
  interactions_XC,
  intervention_type,
  effect_measure,
  separation_method,
  mcsim
)
}
\arguments{
\item{dat}{A \code{data.frame} containing the dataset for analysis.}

\item{ind}{A \code{vector} of indices specifying the subset of \code{dat} to use for the analysis.
Defaults to all rows of \code{dat}. This parameter is particularly useful when using this function within the
\code{boot()} function from the \code{boot} package, as it enables resampling by specifying subsets of the data.}

\item{first}{An \code{integer} specifying the index of the first mediator of interest in the combined list of
intermediate confounders and mediators.}

\item{K}{An \code{integer} specifying the total number of mediators and intermediate confounders.
Mediators are considered sequentially based on their order.}

\item{fam_type}{A \code{character} string specifying the family type for modeling. Options typically include
\code{"gaussian"} for continuous variables or \code{"binomial"} for binary variables.}

\item{mediators}{A \code{character} vector including the variable names for mediators (including intermediate
confounders).}

\item{interactions_XC}{A \code{character} string specifying the two-way interactions amongst exposure and baseline confounders
to include in the regression models in the estimation procedure. The default value, \code{"all"},
includes all two-way exposure-confounder interactions but excludes confounder-confounder interactions.
Specify \code{"none"} to exclude all two-way interactions amongst exposure and baseline confounders.}

\item{intervention_type}{A \code{character} string indicating the type of interventional effect to be estimated.
Options include:
\itemize{
\item \code{"all"} (default): Estimates all types of interventional indirect effects.
\item \code{"shift_all"}: Estimates an interventional indirect effect mapped to a target trial assessing the impact of shifting the joint distribution of all
mediators in the exposed to match the corresponding distribution in the unexposed.
\item \code{"shift_k"}: Estimates an interventional indirect effect mapped to a target trial assessing the impact of shifting the distribution of a specific
mediator (\code{k}) in the exposed to match the corresponding distribution in the unexposed.
\item \code{"shift_k_order"}: Estimates an interventional indirect effect mapped to a target trial assessing the impact of shifting the distribution of a
specific mediator (\code{k}) in the exposed to match the corresponding distribution in the unexposed while accounting for the flow-on
effects on causally descendant mediators.
}}

\item{effect_measure}{A \code{character} string specifying the effect measure to calculate.
For a continuous outcome, only \code{"Diff"} (difference in means) is allowed.
For a binary outcome, must be either \code{"RD"} (risk difference) or \code{"RR"} (risk ratio).
If not specified, defaults to \code{"Diff"} for continuous outcomes and \code{"RD"} for binary outcomes.
Only one effect measure can be specified at a time.}

\item{separation_method}{Method to handle separation, only relevant for binomial (binary outcome) models.
Options are \code{"brglm"} (fits a bias-reduced GLM using the \code{brglm2} package)
or \code{"discard"} (if separation is detected using the \code{detectseparation} package, the function returns \code{NA}
and the bootstrap sample is discarded).}

\item{mcsim}{An \code{integer} specifying the number of Monte Carlo simulations to perform.}
}
\description{
This function performs the actual causal mediation analysis to estimate interventional effects mapped to a
hypothetical target trial.
}
\keyword{internal}
