% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_exposure}
\alias{set_exposure}
\title{Set Exposure Column to a Specific Value and Convert to Factor}
\usage{
set_exposure(data, column_name, exp_val)
}
\arguments{
\item{data}{A \code{data.table} object.}

\item{column_name}{A character string specifying the name of the column to modify.}

\item{exp_val}{The exposure value to assign to the column.}
}
\description{
Set Exposure Column to a Specific Value and Convert to Factor
}
\keyword{internal}
