% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{summary.medRCT}
\alias{summary.medRCT}
\title{Summarise Results from Mediation Analysis}
\usage{
\method{summary}{medRCT}(object, ...)
}
\arguments{
\item{object}{An object of class \code{medRCT}.}

\item{...}{Additional arguments to customise the summary output.}
}
\description{
Provides a summary of the results obtained from a mediation analysis conducted
using the \code{medRCT} function.
}
