% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medRCT_shiny.R
\name{gen_formula_shiny}
\alias{gen_formula_shiny}
\title{Generate Model Formulas for Shiny-Based Model Diagnostics}
\usage{
gen_formula_shiny(
  mediators = mediators,
  exposure = exposure,
  k,
  first = NULL,
  MM = NULL,
  K = NULL,
  interactions_XC,
  include_all = FALSE,
  marginal = FALSE
)
}
\arguments{
\item{mediators}{A \code{character} vector specifying the names of mediator variables in the dataset. The mediators
can be specified as either binary or continuous. When estimating the effect type \code{"shift_k_order"},
the order of mediators in the vector is important, as it determines the causal sequence of mediators.}

\item{exposure}{A \code{character} string specifying the name of the exposure variable in the dataset.
The exposure variable can be categorical, with \code{0} explicitly denoting the unexposed (or control) group.
Other values represent different exposure categories.}

\item{k}{An integer specifying the index of the mediator for which the formula is being generated.}

\item{first}{An integer (optional) specifying the index of the first mediator. Defaults to \code{NULL},
indicating that the function will generate formulas without explicitly considering this parameter.}

\item{MM}{An integer (optional) specifying the index of the mediator whose distribution will be shifted.
Defaults to \code{NULL}, indicating that the function will generate formulas without explicitly considering
this parameter.}

\item{K}{An integer (optional) specifying the total number of mediators and intermediate confounders.
Defaults to \code{NULL}, indicating that the function will generate formulas without explicitly considering this
parameter.}

\item{interactions_XC}{A \code{character} string specifying the exposure-confounder or confounder-confounder
interaction terms to include in the regression models for confounder adjustment. The default value, \code{"all"},
includes all two-way exposure-confounder interactions but excludes confounder-confounder interactions.
Specify \code{"none"} to exclude all two-way exposure-confounder and confounder-confounder interactions.}

\item{include_all}{Logical.}

\item{marginal}{Logical. If \code{TRUE}, estimating marginals under \code{X=0}.}
}
\description{
This function generates model formulas using the original variable names as they appear in the dataset.
}
\keyword{internal}
