% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_proc.R
\name{joint_X_nonzero}
\alias{joint_X_nonzero}
\title{Estimation of Joint Distribution of All Other Mediators under Exposed}
\usage{
joint_X_nonzero(
  MM,
  k,
  first,
  K,
  data,
  dat2,
  fam_type,
  mediators,
  interactions_XC,
  lnzero,
  separation_method,
  n,
  index
)
}
\arguments{
\item{MM}{An integer specifying the index of the mediator whose distribution
will be shifted.}

\item{k}{An integer specifying the index of the current mediator being processed.}

\item{first}{An integer specifying the index of the first mediator.}

\item{K}{An integer specifying the total number of mediators and intermediate confounders.}

\item{data}{A \code{data.frame} containing the dataset for analysis. The dataset is
used to fit the model for the mediators.}

\item{dat2}{A \code{data.frame} used to perform counterfactual predictions and
random draws.}

\item{fam_type}{A \code{character} string specifying the family type for modeling.
Options typically include \code{"gaussian"} for continuous variables or \code{"binomial"}
for binary variables.}

\item{mediators}{A \code{character} vector including the variable names for mediators
(including intermediate confounders).}

\item{interactions_XC}{A \code{character} string specifying the two-way interactions amongst exposure and baseline confounders
to include in the regression models in the estimation procedure.}

\item{lnzero}{A numeric vector specifying the non-zero levels of the exposure.}

\item{separation_method}{Method to handle separation, only relevant for binomial (binary outcome) models.
Options are \code{"brglm"} (fits a bias-reduced GLM using the \code{brglm2} package)
or \code{"discard"} (if separation is detected using the \code{detectseparation} package, the function returns \code{NA}
and the bootstrap sample is discarded).}

\item{n}{An integer specifying the number of observations for \code{dat2}.}

\item{index}{An integer vector specifying the indices of all mediators, excluding the mediator
specified by \code{MM}.}
}
\description{
Estimation of Joint Distribution of All Other Mediators under Exposed
}
\keyword{internal}
