% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LSAC.R
\docType{data}
\name{LSACdata}
\alias{LSACdata}
\title{The Longitudinal Study of Australian Children data}
\format{
A data frame containing following variables from a longitudinal cohort study.
\describe{
\item{child_sex:}{ the binary}
\item{child_atsi}{ the binary}
\item{mat_cob}{ the binary}
\item{mat_engl}{ the binary}
\item{mat_age}{ the binary}
\item{sep}{ the binary}
\item{fam_stress}{ the binary}
\item{parent_mh}{ the binary}
\item{preschool_att}{ the binary}
\item{child_mh}{ the binary}
}
}
\usage{
LSACdata
}
\description{
Simulated data based on the Longitudinal Study of Australian Children containing 5,107 rows and
10 columns with no missing values.
}
\keyword{datasets}
