% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_fun.R
\name{medRCT.fun}
\alias{medRCT.fun}
\title{Causal Mediation Analysis for Estimating Interventional Effects}
\usage{
medRCT.fun(
  dat,
  ind = 1:nrow(dat),
  first = first,
  K = K,
  fam_type = fam_type,
  interactions_XC = interactions_XC,
  intervention_type = intervention_type,
  mcsim
)
}
\arguments{
\item{dat}{A \code{data.frame} containing the dataset for analysis.}

\item{ind}{A \code{vector} of indices specifying the subset of \code{dat} to use for the analysis.
Defaults to all rows of \code{dat}. This parameter is particularly useful when using this function within the
\code{boot()} function from the \code{boot} package, as it enables resampling by specifying subsets of the data.}

\item{first}{An \code{integer} specifying the index of the first mediator of interest in the combined list of
intermediate confounders and mediators.}

\item{K}{An \code{integer} specifying the total number of mediators and intermediate confounders.
Mediators are considered sequentially based on their order.}

\item{fam_type}{A \code{character} string specifying the family type for modeling. Options typically include
\code{"gaussian"} for continuous variables or \code{"binomial"} for binary variables.}

\item{interactions_XC}{A \code{character} string specifying the exposure-confounder or confounder-confounder
interaction terms to include in the regression models for confounder adjustment. The default value, \code{"all"},
includes all two-way exposure-confounder interactions but excludes confounder-confounder interactions.
Specify \code{"none"} to exclude all two-way exposure-confounder and confounder-confounder interactions.}

\item{intervention_type}{A \code{character} string indicating the type of interventional effect to be estimated.
Options include:
\itemize{
\item \code{"all"} (default): Estimates all types of interventional indirect effects.
\item \code{"shift_all"}: Estimates the interventional indirect effect of shifting the joint distribution of all
mediators in the exposed to match the level in the unexposed.
\item \code{"shift_k"}: Estimates the interventional indirect effect of shifting the distribution of a specific
mediator (\code{k}) in the exposed to match the level in the unexposed.
\item \code{"shift_k_order"}: Estimates the interventional indirect effect of shifting the distribution of a
specific mediator (\code{k}) in the exposed to match the level in the unexposed while accounting for the flow-on
effects on its causal descendent mediators.
}}

\item{mcsim}{An \code{integer} specifying the number of Monte Carlo simulations to perform.}
}
\description{
This function performs the actual causal mediation analysis to estimate interventional effects mapped to a
hypothetical target trial.
}
