% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medRCT.R
\name{medRCT.fun}
\alias{medRCT.fun}
\title{Causal mediation analysis for estimating the interventional effect mapped to a target trial}
\usage{
medRCT.fun(
  dat,
  ind = 1:nrow(dat),
  first = first,
  K = K,
  Xconfsint = Xconfsint,
  effect_type = effect_type,
  mcsim
)
}
\arguments{
\item{dat}{A data.frame with the data for analysis}

\item{ind}{A vector of indices that define the sample from dat on which to conduct the analysis. Defaults to all rows of the data.
This facilitates the use of this function within the boot() function from the boot package}

\item{first}{index of first mediator of interest after combining intermediate confounders with mediators}

\item{K}{the number of mediators}

\item{Xconfsint}{a character string specifying the exposure-confounder or confounder-confounder interaction terms
to include in all regression models in the procedure. Defaults to include all two-way exposure-confounder interactions
and no confounder-confounder interactions.}

\item{effect_type}{a character string indicating the type of the interventional effect to be estimated.
Can be 'all', 'E_all', 'E_k', 'E_prime'. Default is 'all'.}

\item{mcsim}{the number of Monte Carlo simulations to conduct}
}
\description{
Causal mediation analysis for estimating the interventional effect mapped to a target trial
}
