% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medRCT.R
\name{medRCT}
\alias{medRCT}
\title{Causal mediation analysis for estimating the interventional effect}
\usage{
medRCT(
  dat,
  exposure,
  outcome,
  mediators,
  int.confounders,
  confounders,
  Xconfsint = paste(paste(rep("X", length(confounders)), confounders, sep = "*"),
    collapse = "+"),
  effect_type = c("all", "E_all", "E_k", "E_prime"),
  mcsim,
  boot_args = list(R = 100, stype = "i"),
  ...
)
}
\arguments{
\item{dat}{a data.frame with the data for analysis}

\item{exposure}{a character string with name of the exposure in the data. The exposure variable can only be binary.}

\item{outcome}{a character string with name of the outcome in the data. The outcome variable con only be binary.}

\item{mediators}{a character vector with the names of the mediators in data. The mediators can only be binary variables.}

\item{int.confounders}{a character vector with the names of the intermediate confounders in data. The intermediate
confounders can only be binary variables.}

\item{confounders}{a character vector with names of the confounders in data, which must be of the required class
(e.g. factor if appropriate)}

\item{Xconfsint}{a character string specifying the exposure-confounder or confounder-confounder interaction terms
to include in all regression models in the procedure. Defaults to include all two-way exposure-confounder interactions
and no confounder-confounder interactions.}

\item{effect_type}{a character string indicating the type of the interventional effect to be estimated.
Can be 'all', 'E_all', 'E_k', 'E_prime'. Default is 'all'.}

\item{mcsim}{the number of Monte Carlo simulations to conduct}

\item{boot_args}{a \code{list} of bootstrapping arguments. \code{R} is the number of bootstrap replicates.
\code{stype} indicates what the second argument of \code{statistics} in the \code{boot} function represents}

\item{...}{other arguments passed to the \code{boot} package.}
}
\description{
'medRCT' is used to estimating the interventional effects that emulate a target trial with any number of multiple
mediators, including some not of primary interest but that are intermediate confounders
}
