% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medRCT.R
\name{medRCT}
\alias{medRCT}
\title{Causal mediation analysis using interventional effects that map to a target trial}
\usage{
medRCT(
  dat,
  exposure,
  outcome,
  mediators,
  intermediate_confs,
  confounders,
  interactions_XC = "all",
  intervention_type = c("all", "shift_all", "shift_k", "shift_k_order"),
  mcsim,
  bootstrap = TRUE,
  boot_args = list(R = 100, stype = "i", ci.type = "norm"),
  ...
)
}
\arguments{
\item{dat}{A \code{data.frame} containing the dataset for analysis. It should include variables for the exposure,
outcome, mediators, confounders, and exposure-induced mediator-outcome confounders specified in the analysis.}

\item{exposure}{A \code{character} string specifying the name of the exposure variable in the dataset.
The exposure variable can be categorical, with \code{0} explicitly denoting the unexposed (or control) group.
Other values represent different exposure categories.}

\item{outcome}{A \code{character} string specifying the name of the outcome variable in the dataset.
The outcome variable can be specified as either binary or continuous.}

\item{mediators}{A \code{character} vector specifying the names of mediator variables in the dataset. The mediators
can be specified as either binary or continuous. When estimating the effect type \code{"shift_k_order"},
the order of mediators in the vector is important, as it determines the causal sequence of mediators.}

\item{intermediate_confs}{A \code{character} vector specifying the names of intermediate confounders in the dataset.
The intermediate confounders can be specified as either binary or continuous. If \code{NULL},
no intermediate confounders are specified, and the natural effect will be estimated.}

\item{confounders}{A \code{character} vector listing the names of baseline confounders.}

\item{interactions_XC}{A \code{character} string specifying the exposure-confounder or confounder-confounder
interaction terms to include in the regression models for confounder adjustment. The default value, \code{"all"},
includes all two-way exposure-confounder interactions but excludes confounder-confounder interactions.
Specify \code{"none"} to exclude all two-way exposure-confounder and confounder-confounder interactions.}

\item{intervention_type}{A \code{character} string indicating the type of interventional effect to be estimated.
Options include:
\itemize{
\item \code{"all"} (default): Estimates all types of interventional indirect effects.
\item \code{"shift_all"}: Estimates the interventional indirect effect of shifting the joint distribution of all
mediators in the exposed to match the level in the unexposed.
\item \code{"shift_k"}: Estimates the interventional indirect effect of shifting the distribution of a specific
mediator (\code{k}) in the exposed to match the level in the unexposed.
\item \code{"shift_k_order"}: Estimates the interventional indirect effect of shifting the distribution of a
specific mediator (\code{k}) in the exposed to match the level in the unexposed while accounting for the flow-on
effects on its causally descendent mediators.
}}

\item{mcsim}{An \code{integer} specifying the number of Monte Carlo simulations to perform.}

\item{bootstrap}{A \code{logical} value indicating whether bootstrapping should be performed. If \code{TRUE}
(default), bootstrapping is conducted using the \code{boot} function from the \code{boot} package.}

\item{boot_args}{A \code{list} of arguments for bootstrapping. The default settings are:
\itemize{
\item \code{R}: Number of bootstrap replicates (default: 100).
\item \code{stype}: Specifies the statistic type passed to the \code{boot} function (default: \code{"i"}).
\item \code{ci.type}: Specifies the type of confidence interval to compute (default: \code{"norm"}).
}}

\item{...}{Additional arguments passed to the \code{boot} function from the \code{boot} package.}
}
\description{
'medRCT' is used to estimate interventional effects that map to a target trial evaluating hypothetical mediator interventions of interest.
It can handle multiple mediators, including some not of primary interest but that are exposure-induced
mediator-outcome confounders.
}
