% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medRCT_shiny.R
\name{medRCT_shiny}
\alias{medRCT_shiny}
\title{Launch Interactive Shiny App for Model Diagnostics}
\usage{
medRCT_shiny(data, ...)
}
\arguments{
\item{data}{A \code{data.frame} containing the dataset for analysis. It should include variables for the exposure,
outcome, mediators, confounders, and exposure-induced mediator-outcome confounders specified in the analysis.}

\item{...}{additional arguments for shiny}
}
\description{
\code{medRCT_shiny} launches a Shiny application to provide model summaries
of all models fitted by the algorithm. The app serves as a user-friendly
interface for diagnosing potential issues with the models and facilitates adjustments
to improve model fit.
}
\examples{
if (interactive()) {
   medRCT_shiny(data=LSACdata)
}
}
